Spread out the components into bits ,for better work enviroment. Added a fecthing functionallity,since the API in osted on the web. Used useEffect to fecth the data with an [] as a dependecy on the useEffect function.
Created a const [loading,setLoading] = useState(true) to set up a conditional statement,wheres after the fecth function the loading is set to false ,thus triggering a render of the pages Tours component.
Used a .map() function to iterate throught the fecthed data in the API that was turned into a json() document.